third_party/blink/web_tests/fast/css/attribute-selector-case-insensitive-expected.txt

Testing exact attribute selector.
PASS window.getComputedStyle(element).borderTopColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderRightColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderBottomColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderLeftColor is 'rgb(0, 128, 0)'
Testing list attribute selector.
PASS window.getComputedStyle(element).borderTopColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderRightColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderBottomColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderLeftColor is 'rgb(0, 128, 0)'
Testing hyphen attribute selector.
PASS window.getComputedStyle(element).borderTopColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderRightColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderBottomColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderLeftColor is 'rgb(0, 128, 0)'
Testing begin attribute selector.
PASS window.getComputedStyle(element).borderTopColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderRightColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderBottomColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderLeftColor is 'rgb(0, 128, 0)'
Testing end attribute selector.
PASS window.getComputedStyle(element).borderTopColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderRightColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderBottomColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderLeftColor is 'rgb(0, 128, 0)'
Testing contains attribute selector.
PASS window.getComputedStyle(element).borderTopColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderRightColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderBottomColor is 'rgb(0, 128, 0)'
PASS window.getComputedStyle(element).borderLeftColor is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE


PASS if all elements borders are all green, not red.
